jsonrpc: convert multi-agent-rotate-authentication-key-admin-script test to forge
